Update
Moved builtin_types to ParserState.py
Changed builtin and std to now require rial: prefix
Removed cache file hash checking to speed up worst case (and if file modification doesn't match then it most likely changed)
Improved compilation speed when cache is disabled
Improved compatibility with Python versions

Added ability to work singlethreaded-ish. This was a giant change that completely overhauled the way the compilation units and scheduling was done. Before when a thread encounters a using statement it would block until that using statement was resolved....Now if theres a sufficient number of modules and none of them could completely be resolved you'd get a deadlock.
So now threads pack it all up into compilation_unit.py(s) and start working on other files in a double-list fashion. As a result the in-operation support for resolving uses from function names was removed (it was dumb anyways honestly). There's still a lot to be refactored and improved though, especially in regard to readability.

Profiler now checks if an event was already recorded and will append the time to the first event